+++ src/openassetio-core/include/openassetio/errors/BatchElementError.hpp
@@ -7,12 +7,13 @@
 #include <string>
 #include <utility>
 
-#include <openassetio/errors/errors.h>
+#include <openassetio/errors/errorCodes.h>
 #include <openassetio/export.h>
 #include <openassetio/typedefs.hpp>
 
 namespace openassetio {
 inline namespace OPENASSETIO_CORE_ABI_VERSION {
+namespace errors {
 /**
  * @name Batch element errors
  *
@@ -153,87 +154,8 @@ static_assert(std::is_move_constructible_v<BatchElementError>);
 static_assert(std::is_move_assignable_v<BatchElementError>);
 
 /**
- * Exception base that ties together a @ref BatchElementError and an
- * index.
- *
- * When thrown from a function, indicates that a particular
- * element has caused an error. The specific element that has errored
- * is indicated by the index attribute, relative to the input container.
- */
-struct OPENASSETIO_CORE_EXPORT BatchElementException : std::runtime_error {
-  BatchElementException(std::size_t idx, BatchElementError err)
-      : std::runtime_error{err.message}, index{idx}, error{std::move(err)} {}
-
-  /**
-   * Index describing which batch element has caused an error.
-   */
-  std::size_t index;
-
-  /**
-   * Object describing the nature of the specific error.
-   */
-  BatchElementError error;
-};
-
-/**
- * Exception equivalent of
- * @ref BatchElementError.ErrorCode.kUnknown
- */
-struct OPENASSETIO_CORE_EXPORT UnknownBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref BatchElementError.ErrorCode.kInvalidEntityReference
- */
-struct OPENASSETIO_CORE_EXPORT InvalidEntityReferenceBatchElementException
-    : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref BatchElementError.ErrorCode.kMalformedEntityReference
- */
-struct OPENASSETIO_CORE_EXPORT MalformedEntityReferenceBatchElementException
-    : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref BatchElementError.ErrorCode.kEntityAccessError
- */
-struct OPENASSETIO_CORE_EXPORT EntityAccessErrorBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref BatchElementError.ErrorCode.kEntityResolutionError
- */
-struct OPENASSETIO_CORE_EXPORT EntityResolutionErrorBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref BatchElementError.ErrorCode.kInvalidPreflightHint
- */
-struct OPENASSETIO_CORE_EXPORT InvalidPreflightHintBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-
-/**
- * Exception equivalent of
- * @ref BatchElementError.ErrorCode.kInvalidTraitSet
- */
-struct OPENASSETIO_CORE_EXPORT InvalidTraitSetBatchElementException : BatchElementException {
-  using BatchElementException::BatchElementException;
-};
-/**
  * @}
  */
+}  // namespace errors
 }  // namespace OPENASSETIO_CORE_ABI_VERSION
 }  // namespace openassetio